JPA DAO

+++ 05-JPA-API/src/br/com/fiap/jpa/dao/impl/BebidaDAOImpl.java
@@ -4,6 +4,8 @@ import javax.persistence.EntityManager;
 
 import br.com.fiap.jpa.dao.BebidaDAO;
 import br.com.fiap.jpa.entity.Bebida;
+import br.com.fiap.jpa.exception.CommitException;
+import br.com.fiap.jpa.exception.NoDataException;
 
 public class BebidaDAOImpl implements BebidaDAO {
 
@@ -16,15 +18,11 @@ public class BebidaDAOImpl implements BebidaDAO {
 	@Override
 	public void cadastrar(Bebida bebida) {
 		em.persist(bebida);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
 	public void atualizar(Bebida bebida) {
 		em.merge(bebida);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
 	}
 
 	@Override
@@ -33,11 +31,25 @@ public class BebidaDAOImpl implements BebidaDAO {
 	}
 
 	@Override
-	public void apagar(int codigo) {
+	public void apagar(int codigo) throws NoDataException {
 		Bebida bebida = pesquisar(codigo);
+		//valida se existe uma bebida para ser removida
+		if (bebida == null) {
+			throw new NoDataException();
+		}
 		em.remove(bebida);
-		em.getTransaction().begin();
-		em.getTransaction().commit();
+	}
+
+	@Override
+	public void commit() throws CommitException {
+		try {
+			em.getTransaction().begin();
+			em.getTransaction().commit();
+		}catch(Exception e) {
+			e.printStackTrace();
+			em.getTransaction().rollback();
+			throw new CommitException();
+		}
 	}
 
 }
